+++ tests/tst/TargetChild.php
@@ -3,7 +3,7 @@
  * @package axy\creator
  */
 
-namespace axy\creator\tests\nstst;
+namespace axy\creator\tests\tst;
 
 class TargetChild extends Target
 {