+++ errors/Disabled.php
@@ -8,6 +8,8 @@ namespace axy\creator\errors;
 
 /**
  * Nested service is disabled
+ *
+ * @link https://github.com/axypro/creator/blob/master/doc/errors.md documentation
  */
 class Disabled extends \axy\errors\Disabled
 {