+++ errors/Error.php
@@ -8,6 +8,8 @@ namespace axy\creator\errors;
 
 /**
  * Basic error of axy\creator
+ *
+ * @link https://github.com/axypro/creator/blob/master/doc/errors.md documentation
  */
 interface Error extends \axy\errors\Error
 {